Text Notes Adjustment
Looking carefully at the transcriptions between Object 12 and 16 yields
one key difference! In Object 16 the <physnumber> tag is closed right
after the number, but in Object 12 it is closed outside of the <note>
tag.
Dito for Object 8 and 31.

Like Joe F saw,
Object 7 is a problem because the <note> tag is within the <catchword>
tag and should be outside of it. Same with Object 24.

Object 12 looked good except for maybe the placement of the <space>
after <physcnumber> and <notes>, I moved <space> in Object 12 to right
next to the <l> tag, to see if that works.