Celeste Seifert update of songsie Copy U.
I will do another review when I have time to see if I missed any <hi rend=“i”> changes, but when I was going through it I decided to compare the XML of Copy U on the archive with the one here on github, but I could not find it listed. Is there a version uploaded on the site, or is this a Copy not yet uploaded?